+++ misc/linear_smoother.cpp
@@ -28,7 +28,7 @@ LinearSmoother::LinearSmoother()
 
 }
 
-void LinearSmoother::set_sample_rate(double _sample_rate)
+void LinearSmoother::set_sample_rate(float _sample_rate)
 {
 	sample_rate = _sample_rate; 
 	sr_cst = 1.0 / sample_rate;
